refactor: make repository package
