+++ src/main/kotlin/com/kroffle/knitting/infra/configuration/AppConfig.kt
@@ -5,10 +5,10 @@ import com.kroffle.knitting.infra.jwt.TokenPublisher
 import com.kroffle.knitting.infra.oauth.GoogleOAuthHelperImpl
 import com.kroffle.knitting.infra.oauth.dto.ClientInfo
 import com.kroffle.knitting.infra.oauth.dto.GoogleOAuthConfig
-import com.kroffle.knitting.infra.persistence.design.DBDesignRepository
-import com.kroffle.knitting.infra.persistence.design.R2dbcDesignRepository
-import com.kroffle.knitting.infra.persistence.knitter.DBKnitterRepository
-import com.kroffle.knitting.infra.persistence.knitter.R2dbcKnitterRepository
+import com.kroffle.knitting.infra.persistence.design.repository.DBDesignRepository
+import com.kroffle.knitting.infra.persistence.design.repository.R2dbcDesignRepository
+import com.kroffle.knitting.infra.persistence.knitter.repository.DBKnitterRepository
+import com.kroffle.knitting.infra.persistence.knitter.repository.R2dbcKnitterRepository
 import com.kroffle.knitting.infra.properties.AuthProperties
 import com.kroffle.knitting.infra.properties.ClientProperties
 import com.kroffle.knitting.infra.properties.SelfProperties
